HACK to get normalized URLs working correctly
Since normalize runs during the build now and not when the page is executed, all of our URLs within our less files are not correct. This is a terrible fix to get our URLs working. Need to open an issue with the main project to get a more permanent fix.